+++ src/cc/h/ctypes.h
@@ -483,7 +483,7 @@ enum quad_type {
 
 typedef struct {
     union   {
-        long        long_values[2];
+        uint32_t    long_values[2];
         int64       long64;
         double      double_value;
         long_double long_double_value;
