+++ src/cc/h/cvars.h
@@ -98,7 +98,7 @@ global  source_loc SrcFileLoc;  /* duplicate of SrcFile->src_line */
 global  source_loc CommentLoc;
 global  int     CurrChar;
 global  DATA_TYPE  ConstType;
-global  unsigned long Constant;
+global  uint32_t   Constant;
 global  uint64  Const64;
 global  int     CLitLength;     /* length of string literal */
 global  FCB     *MainSrcFile;   /* primary source file being compiled */
